src/app/prototypes/testin-events/server/pyproject.toml
[project]
name = "columbia-events-mcp"
version = "0.1.0"
description = "MCP server for fetching and managing Columbia University events"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dedalus-labs>=0.2.0",
    "dedalus-mcp>=0.7.0",
    "python-dotenv>=1.2.1",
    "httpx>=0.28.0",
]
